requirements.txt
aiobotocore==3.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aioredis==2.0.1
aiosignal==1.4.0
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
authx==1.4.3
bcrypt==5.0.0
beanie==2.0.1
bidict==0.23.1
billiard==4.2.4
botocore==1.42.5
celery==5.6.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.119.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jmespath==1.0.1
kombu==5.6.1
lazy-model==0.4.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pillow_heif==1.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.1
pydantic-settings==2.11.0
pydantic_core==2.41.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.15.0
pytz==2025.2
redis==7.1.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
